Move globbing utilities to glob module
